+++ src/main/java/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/RobotKeywordCallNameReference.java
@@ -89,7 +89,11 @@ public class RobotKeywordCallNameReference extends PsiPolyVariantReferenceBase<R
                                                          .filter(keywordFile -> keywordFile.getImportType() != ImportType.VARIABLES)
                                                          .map(KeywordFile::getVirtualFile)
                                                          .collect(Collectors.toSet());
-        importedFiles.add(robotFile.getViewProvider().getVirtualFile());
+        VirtualFile virtualFile = robotFile.getVirtualFile();
+        if (virtualFile == null) {
+            virtualFile = robotFile.getOriginalFile().getVirtualFile();
+        }
+        importedFiles.add(virtualFile);
         return importedFiles;
     }
 }