+++ src/main/kotlin/dev/xeonkryptos/xeonrobotframeworkplugin/psi/ref/ResourceFileImportFinder.kt
@@ -29,7 +29,7 @@ class ResourceFileImportFinder(private val project: Project) {
         }
 
         // First, try to resolve the resource file path relative to the location of the importing file.
-        val virtualSourceFile = sourceFile.viewProvider.virtualFile
+        val virtualSourceFile = sourceFile.virtualFile ?: sourceFile.originalFile.virtualFile ?: return null
         val relevantDir = if (virtualSourceFile.isDirectory) virtualSourceFile else virtualSourceFile.parent
         relevantDir?.findFileByRelativePath(resourceFileValue)?.let {
             return PsiManager.getInstance(project).findFile(it)?.takeIf { psiFile -> expectedFileType == null || psiFile.fileType === expectedFileType }